requirements.txt
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
javaobj-py3==0.4.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
pyjks==20.0.0
twofish==0.3.0
typer==0.9.0
typing_extensions==4.8.0
